Fix circle import in srg_utils
