+++ utils/srg_utils/yaml.py
@@ -1,9 +1,9 @@
 import os
 import re
+from pathlib import Path
 
 from ssg.rule_yaml import find_section_lines, get_yaml_contents
-from ssg.utils import read_file_list
-from utils.srg_utils import create_output, write_output
+from ssg.utils import read_file_list, mkdir_p
 
 
 def add_replacement_to_result(replacement, result):
@@ -62,3 +62,24 @@ def replace_yaml_key(key: str, replacement: str, rule_dir: dict) -> None:
 def update_row(changed: str, current: str, rule_dir_json: dict, section: str) -> None:
     if changed != current and changed:
         replace_yaml_section(section, changed, rule_dir_json)
+
+
+def create_output(rule_dir: str) -> str:
+    path_dir_parent = os.path.join(rule_dir, "policy")
+    mkdir_p(path_dir_parent)
+    path_dir = os.path.join(path_dir_parent, "stig")
+    mkdir_p(path_dir)
+    path = os.path.join(path_dir, 'shared.yml')
+    Path(path).touch()
+    return path
+
+
+def write_output(path: str, result: tuple) -> None:
+    with open(path, 'w') as f:
+        first_time = True
+        for line in result:
+            if first_time:
+                first_time = False
+                line = re.sub(r'^\n', '', result[0])
+            f.write(line.rstrip())
+            f.write('\n')
